Clean user input to only select the entered digits.
When people made mistakes when entering the number it still will
validate the token properly. Besides it will print a validation error
inmediatly if the token is wrong.
We can also be more comfortable sanitizing the user input like this.
I don't want to validate the length of the string because it will break
the script if the API token length changes in the future.